requirements.txt
gltflib
matplotlib
numpy
Pillow
scipy
tqdm
transformers
trimesh